refactor(council): pay for Phases 1-2 by extracting data carriers
check_source_size_budget refused the phase at +122 lines, correctly: orchestrator.ts sits
~750 lines above the 1,500-line ceiling, so any net growth in it is refused. Raising the
baseline is what that gate's own message calls a defect rather than a fix.

Paid entirely out of PURE MOVES into ai_council/orchestrator_results.ts:
  - PeerReviewResult + ConsensusResult   -98
  - RunPeerReviewOptions + RunConsensusScoringOptions  -19
  - CouncilQuestion                      -11

All five are data carriers with no logic and no orchestrator-internal dependencies, so
moving them changes no behaviour. orchestrator.ts re-exports every name, so no import path
anywhere changes — 995 council tests pass unchanged.

The options interfaces needed a TYPE-ONLY back-import of OnOverrunCallback from
orchestrator.ts, which also imports types from the new module. That is legal and
runtime-free because both directions are erased, and the alternative was leaving 19 lines
behind for no design reason.

Extracting data carriers rather than shaving the new comments was deliberate: those
comments explain why a per-reviewer map exists and why the re-ask is bounded at one, and
compressing them buys the same green from worse code.

Total excess 18,574 → 18,572, baseline set to the EXACT live total —
check_source_size_budget.test.ts asserts baseline == tree total, so a baseline left one
line loose is a red test.
